requirements.txt
# ===========================
# CORE DEPENDENCIES
# ===========================
openai>=1.12.0
streamlit>=1.31.0
python-dotenv>=1.0.0

# ===========================
# DOCUMENT PROCESSING
# ===========================
pypdf>=3.17.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
chardet>=5.2.0
unidecode>=1.3.0
lxml>=4.9.0

# ===========================
# NLP AND EMBEDDINGS
# ===========================
tiktoken>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.3.0
spacy>=3.7.0

# ===========================
# IMPORTANT: Version constraints for compatibility
# ===========================
numpy<2.0.0  # CRITICAL: TensorFlow doesn't support NumPy 2.x yet
torch>=2.0.0,<2.2.0  # Compatible version
transformers>=4.36.0,<4.38.0  # Compatible with current torch
tensorflow>=2.13.0,<2.16.0  # If needed by transformers
huggingface-hub>=0.19.0

# ===========================
# VECTOR STORES
# ===========================
chromadb>=0.4.22
faiss-cpu>=1.7.4

# ===========================
# ML AND DATA PROCESSING
# ===========================
# numpy<2.0.0  # Already specified above
scikit-learn>=1.3.0
scipy>=1.10.0
tenacity>=8.2.0

# ===========================
# VISUALIZATION
# ===========================
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0
kaleido>=0.2.1

# ===========================
# LOGGING AND UTILITIES
# ===========================
colorlog>=6.8.0
tqdm>=4.65.0
python-json-logger>=2.0.7

# ===========================
# TESTING
# ===========================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0

# ===========================
# ASYNC SUPPORT
# ===========================
aiofiles>=23.2.0
aiohttp>=3.9.0

# ===========================
# ADDITIONAL UTILITIES
# ===========================
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
urllib3>=2.0.0
pyyaml>=6.0.1
python-multipart>=0.0.6

# ===========================
# DATABASE SUPPORT
# ===========================
sqlalchemy>=2.0.0
duckdb>=0.9.0

# ===========================
# API AND WEB (OPTIONAL)
# ===========================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0

# ===========================
# DEVELOPMENT TOOLS (OPTIONAL)
# ===========================
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
isort>=5.12.0
autopep8>=2.0.0
pylint>=3.0.0

# ===========================
# JUPYTER/NOTEBOOK SUPPORT
# ===========================
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
jupyterlab>=4.0.0